CLN: move get_networkx_graph function
